COZMO-10630 More creative solution for Unity player creation (#3942)
* COZMO-10630 More creative solution for Unity player creation

This change attempts to kill two birds with one stone:
- Creating two unity player instances to fix the webview issue in
  Unity 5.6 is probably doing lots of bad things we should avoid
  if possible
- The startup crash on old devices since moving to Unity 5.6

I think these changes are potentially related, now that OD is
moving to 5.6 and hasn't had the same crashing problems. It seems
like a reasonable assumption that since we're creating two
UnityPlayer instances in Cozmo (to fix the webview problem), that
both instances are likely causing a bitmap allocation that's
pushing us over the edge of what memory we have available.

This change reduces the number of UnityPlayer instances created to
one by removing UnityPlayerActivity from our custom Activity's
inheritance hierarchy, and replacing it with a custom class that's
a carbon copy of UnityPlayerActivity (copied straight from their
source code) but instantiates a CUnityPlayer (the webview fix)
instead of a normal UnityPlayer. This should hopefully prevent us
from making multiple bitmap allocations and thus fix the OOM
crash we get on bootup on some devices.