COZMO-10510 Cozmo fails to dock with cube sometimes (#3956)
* COZMO-10510 Cozmo fails to dock with cube sometimes

In order to pickup a block, CodeLab is doing an AlignWithObject and then a MoveLiftToHeight. The AlignWithObject action is not as precise as a PickupObject action which is what is causing the docking failures.

This is fixed by adding a new DockAction, DA_ALIGN_SPECIAL. This will end up doing the same things, firmware wise, as pickup does. The PickAndPlaceController functions as if the action is DA_ALIGN while the DockingController functions as if it is DA_PICKUP_LOW.

* Auto committing firmware from al/codeLabAlignWithObject