COZMO-10532 Code Lab drop downs in toolbar pop down instead of up (#3991)
Fix bug where the drive distance dropdowns pop down on phones when the blocks are in the toobar. The bug was due to the fact that the num pad was too big to be shown above the block, so the Scratch code would default to displaying it below the block, almost completely out of view off the bottom of the screen. I've modified the code to have smaller num pad buttons (44pxx44px, Apple's recommended minimum) for phone only and other minor changes so num pad fits above block in toolbar on phone.

When the block is not in the toolbar, the block will need to be in the top 1/3rd of the screen for the num pad to fully display.