COZMO-10915, COZMO-10940: Adding Needs Decay/Boost Stub (#3997)
Main objective with the stubbing was to prove out the "LatestEngine" vs "CurrentDisplay" paradigm in NeedsStateManager. Seems to be working well so far. Hopefully continues to be okay with real clad messages, which is the next step.
Used the "values", "brackets" and "partsBroken" data from NeedsState.
Added stub NeedsPlayModal, NeedsEnergyModal, and NeedsRepairModal. Based on mocks/flow these may need to be promoted to Views in the future? Not sure. Transitioning is going to be tricky. Likely to completely change.