COZMO-10817 Beef ID update for 1.5 robots (#3995)
* COZMO-10817 Beef ID update for 1.5 robots

The head/body beef IDs did not contain information that allowed us to differentiate between 1.0, 1.5 white, and 1.5 collectors edition. The bodyColor that is written at the factory will be used to differentiate between different types of robots.

We did not want to modify the length of the IDs so we had to split one of the existing fields in half. Since the hardware version was using 4 bytes (8 characters in the ID string) and we will probably never have anywhere close to 65535 hardware version, so 2 bytes will be sufficient. This means we have 2 bytes for the bodyColor (also plenty) so the new format of the beef body id is bodyColor (2 bytes), hardware version (2 bytes), and bodyID/body serial number(4 bytes).