COZMO-10852 Code Lab: Fix workspace-panning block-disappearing bug (#4018)
Fixes bug where after a block has been dragged off the screen, when workspace is panned, block partially or completely disappears. Fix is in normalize.css.

For dock-with-cube, turn on usage of planner so that docking can be more successful.

Update icons from Seth: wait for face, and wait for happy face, wait for sad face and dock with cube.