COZMO-9313 reset to home pose in codelab (#4042)
Added events from Scratch VM to handle start and stop of scripts

CodeLab triggers a reset-to-home pose after 3 seconds (after a program stops), or immediately at program start, if already in pose it completes immediately.
CodeLab keeps a queue of scratch requests whilst getting robot into the home pose
CodeLab keeps track of whether a program is running, or when green flag and stop blocks are clicked.
Leaving CodeLab now cancels all actions that were in progress, and triggers the CodeLabExit anim
Setting head or lift height now correctly determines difference vs an epsilon, and skips if close enough
Actions for head and lift that would require a promise are forced to do a 0.01 second wait action to ensure that the return is on a callback and not in the same method (otherwise promise isn’t setup yet).

Added WaitAction wrapper
Added note next to kIdealBlockViewHeadValue that they are angle factors, not angles, after having to debug why they were unexpected values.

CodeLab disables all driving animations
Support for specifying numRetries on actions, and interface added for AlignWithObject to use it. CodeLab now requests 2 retries for AlignWithObject

Added a (slightly hacky) build.sh script to make it easier to build Scratch vm in-place and nuke undesired changes.
By default it caches node_modules (but runs npm update to hopefully make it the same as a fresh npm install) as that reduces the time considerably.