COZMO-10751 Crash in BodyLightComponent (#4019)
I think this is the causes of this crash...
We have a list of shared_ptrs and we give out objects that contain both an iterator to an element in the list as well as a weak_ptr to the actual value of the element. We are also internally storing another shared_ptr to one of the elements in the list. This means we could potentially have two shared_ptrs and one weak_ptr all pointing to the same thing. When we go to remove an element from this list of shared_ptrs, we are checking whether or not the weak_ptr is expired before removing. This is an issue if the element has already been removed from the list but we still have the internally stored shared_ptr to it. This means in the object we have given out (with the weak_ptr and iterator), the weak_ptr is still valid but the iterator is not. Trying to erase that iterator results in a crash.

This is fixed by switching the internally stored shared_ptr to a weak_ptr