COZMO-11119 Cancel code lab robot callbacks on stop (#4093)
Stopping programs while playing animations was breaking go-to-home because the go-to-home action would be cancelled the next frame when the animation action callback happened and triggered the reset-to-neutral-face anim.
Fixed by cancelling all callbacks when releasing the blocks. This could mean that an in-progress animation wasn’t able to revert to neutral, so made sure that go-to-home always resets to neutral face.
Removed spurious double entry in CodeLabExit animation trigger.