COZMO-11236 README.md - Update Android SDK/NDK section
Explain to the developer that they don't really have to do anything
unless they care.  If the developer just needs to build Cozmo for
Android, it will all be taken care of automatically.  However, if they
need to debug and troubleshoot on a regular basis, show them what to add
to their `~/.bash_profile` file to set the appropriate environment
variables.