COZMO-11311 Use volatile pointer for CrashTheApp
We want our Anki::Util::CrashTheApp function to try and write the
value of 42 to address 0 in order to generate a seg fault (SIGSEGV).
Breakpad will then catch this seg fault and write out a crash report.

Unfortunately, when doing a release build, the compiler translated our C
code into an ARM/Thumb assembly instruction of "udf" (undefined).  This
is probably because writing to a null pointer is undefined in C.

```
007fd4fc <_ZN4Anki4Util11CrashTheAppEP22ConsoleFunctionContext>:
  7fd4fc:       defe            udf     #254    ; 0xfe
```

This still resulted in a crash and Breakpad writing a crash report, but
not what we would expect.

To fix it, we need to make the pointer volatile to tell the compiler NOT
to optimize this code away.  Doing so produced the expected assembly.

```
008013a4 <_ZN4Anki4Util11CrashTheAppEP22ConsoleFunctionContext>:
  8013a4:       2000            movs    r0, #0
  8013a6:       212a            movs    r1, #42 ; 0x2a
  8013a8:       6001            str     r1, [r0, #0]
  8013aa:       4770            bx      lr
```