COZMO-11245 Code Lab challenge bookmark (#4138)
During a Code Lab game session, persist the page in the challenges that the user last viewed in a Unity variable, _ChallengeBookmark.

We considered saving this on the JavaScript side, but since the challenges are in an iframe, we we unable to find a way to tell the parent page what slide number it stopped on. None of the cross-frame techniques work when the files are loaded from disk and not a domain. Local storage also doesn't work for files loaded from disk.

Also fix prefetch calls to refer to correctly-named files.

Code from Adam and me.