COZMO-11304 Don’t Cancel other anim actions on neutral (#4183)
Previously, clicking on 2 animations in the toolbox in quick succession would previously stop both playing, because the 2nd animation action cancels the 1st, and on the 1st one completing (with success=False) on the next frame, it would then queue a neutral face trigger which would cancel the 2nd animation action.
We now only issue the neutral face anim trigger on successful animation completion.
To ensure that the neutral face will still occur eventually, we now force the ResetToHome script to wait for the neutral trigger (it should only take a couple of engine ticks) instead of just issuing it and hoping. To avoid unnecessary waiting for this, we only wait if one of the other (longer) actions was already queued or if we played an animation since last resetting to home.
Dock with cube had a similar issue where if you press another action whilst Cozmo is driving to the cube it would fail, as the failure case of docking to cube would trigger a failure animation - this is worked-around by queuing that failure in parallel so it won’t clobber any pending actions. There’s probably still an issue if you interrupt dock-with-cube with an animation (as parallel animations can abort existing animations), but that’s rarer and would require us to detect the difference between action failure and interruption from another Scratch block.