COZMO-11147 Auto-build il2cpp code
These changes cause il2cpp code generated from Unity to be automatically
compiled and bundled into static libs. This is implemented using a
combination of cmake + ninja, removing the requirement to explicitly
add il2cpp generated files to Xcode projects.

In the future, we might want to have a way to build these files in
Xcode for debugging purposes. However, this should unblock PRs that
are failing due to merges that cause a change in the size of our
csharp assemblies.

Details:

* Write ANKI_BUILD_UNITY_HOME to ios.xcconfig

We need to variable in order to find the il2cpp library includes that
ship inside of the Unity distribution.

* Create xcode project to build unity

The unity build is now encapulated inside its own xcode project.
There are 2 steps:

1. build unity if necessary (based on check in unity.py)

2. if il2cpp, then use cmake + ninja to compile il2cpp source
   files into a static library

These steps will be used by the CozmoUnity_IOS to build il2cpp without
having to rely on explicitly defining the generated source files in Xcode.

* Update CozmoUnity proj to use unity build project

- Delete Native/* source files (il2cpp generated files)

- Link against static lib built from generated il2cpp source files

- Add unitybuild project to workspace

* Add script to install cmake dep

We depend on at least having cmake 3.6.0. Developers or build agents
might need/have other versions installed. cmake.py will download a
sandboxed version of cmake 3.8.1 into ~/.anki/cmake, so that it will
not change unless we need to change it.

configure.py will auto-install cmake if necessary and pass the path
to the executable to Xcode, which uses it during the build.

* Modify csharp "xcode fixup" script to exclude Native/ dir that
  contains source files generated by il2cpp