COZMO-11147 force_load libunity-generated.a (#4289)
This library contains all of the code generated by il2cpp. The symbols are intended to be
compiled directly into the app. Since we are creating a static lib, we need to force load
the static lib to ensure that all of the symbols are available to the app.

In addition, these changes reduce the code size of the app:

- Enable proper code-strip settings for iOS project

Enable `Deployment Postprocessing` and `Use Separate Strip` on the
executable in order to properly remove dead code and debug symbols.

- Always build il2cpp generated code as optimized with `-Os`

The initial build config used `-O0` all the time, but we don't need to debug this code.
This reduces the size of the raw static lib by about 30 MB.