COZMO-11147 Fix debug ios builds (#4300)
The unity-generated target needs to set "Build Active Arch Only" to NO for
all configurations.  This ensures that Xcode will always consider the lib product
a dependency regardless of the arch that is currently being built. Without this
step, debug builds were failing because the unity build was not running before
the scripts in the CozmoUnity_IOS target.