COZMO-11748 Code Lab: Remove save on close button to fix crash #4322
We currently save every 3 seconds. In order to save when the user exits (in case any changes were made since the last 3-second save), we will also later add a save when the user views the "are you sure you want to leave the code lab workspace?" dialog. But saving here has injected an intermittent crash on iOS and Android, so removing this.