COZMO-11204 Cozmo can not roll off 1.5 chargers (#4239)
* COZMO-11204 Cozmo can not roll off 1.5 chargers

Due to the changes to the 1.5 charger, Cozmo now detects a cliff right as he is driving off it. In order to prevent stopping due to this cliff, the cliff detect threshold is lowered until Cozmo has driven 50mm or while he is on the charger platform. Once he has driven 50 mm or is no longer on the charger platform, the threshold is set back to default value.

This may also fix COZMO-5398