COZMO-11832 Backpack lights stay on when sdk program ends (#4355)
The backpack lights turning off was relying on the bug causing COZMO-10751 so when that got fixed this broke. Fixing COZMO-10571 made it so that all backpackLightConfigs now properly become invalid when clearing configs on SDK program end. A special case has been added for null configs in BodyLightComponent::Update() to turn off the backpack lights