COZMO-11552 Add separate gyp configs for cozmo2 targets (#4327)
This changeset adds cozmoEngine2.gyp, cozmoGame2.gyp and modifies configure
scripts to accept a new `--engine-v2` or `-2` param to generate "Cozmov2" workspaces
for xcode that include only the CozmoEngine2 targets.  For android builds, this flag
causes libcozmoEngine2.so to be built.

This approach provides a clear separate between CozmoEngine targets and paves the
way for migrating cozmov2 builds to a separate build system without affecting
cozmov1.

change by @kevinatanki
* Made it so that the Webots supervisor needs to be instantiated outside of libcozmoEngine2_sim so that other controllers that link to it don’t complain about multiple Robot instances.

* Generate correct engine target for Android

Mac and iOS platforms need to generate Xcode projects for both
cozmoEngine & cozmoEngine2 because of the way we call
configure_engine to create them.  However, for android & linux,
we should only generate the ninja build config for the specified
target.

* Use new workspace for v2 tests

When `--cozmo_version 2` is passed as a param, we need to use the
v2 workspace.

In addition, the build agents need to run this configure step if the
option is available.