COZMO-11481 Code Lab: add exit confirmation dialog (#4390)
Upon closing the save/load UI, an exit confirmation dialog will appear.

This change refactors the dialog code into one location so that it is shared with other dialogs. Currently the only other dialog is that for confirming the deletion of a user project.

One effect of this change is that the order of the "cancel" and "confirm" buttons have changed in the delete user project dialog, to match the order of the "quit activity" dialog buttons in the rest of the app.

Code from Adam/CodeRefactory.