COZMO-11890 Cozmo fails to align with cube (#4395)
* COZMO-11890 Cozmo fails to align with cube

The root cause of this issue was that we were considering lift movement to allow the robot to localize to objects. This means that whenever the lift moved we would constantly localize to an observed objects which was causing the robot's pose to go nuts. These pose jumps were impacting docking, most notable aligning to the cube after moving the lift up.

* Fixing localization unit tests