COZMO-12059 Fix single repair action that fixes more than one part (#4426)
Previously, a user could repair ONE part in the UI, and sometimes TWO parts
would be repaired.  This was because the delta for repair actions was sometimes
large enough that it would cross TWO thresholds of broken parts.  This change
keeps it within the 'broken part thresholds'.  AND another part of this ticket is
to make it so that if there are no more broken parts after a repair, set the
Repair need level to 100% no matter what.

- In NeedsManager::RegisterNeedsActionCompleted(), the changes are simple
  but stuff got moved around also:
- Remove redundant switch statement (oddly, the exact same switch statement
  one right below the other...I thought I had fixed that at some point...)
- Move that switch statement to happen BEFORE the delta application
- Add a check in there that says if there are no more broken parts, set Repair
  level to 100%
- In NeedsState::ApplyDelta (this is the main change), if the delta is for Repair
  and we're going up, clamp the ending Repair need level to make sure it
  doesn't go too high, and that it also goes high enough.