COZMO-11328 Code Lab: Save upon exiting workspace (#4430)
When the user taps the close button to exit the workspace, save the user project. Note that the user project is also saved once every 3 seconds if there have been edits.

When the Cozmo app is backgrounded or the robot disconnects, the CodeLab mini-game is torn down without user interaction, and we don't currently have code that is saving the user project at that time other than the every-3-second save.