COZMO-12235 Cozmo stops interacting with current cube and picks a dif… (#4509)
* COZMO-12235 Cozmo stops interacting with current cube and picks a different one

Due to recent docking changes, we are constantly updating the dock object's pose, even while moving. This results in somewhat inaccurate poses which occasionally will not be considered resting flat. Any behaviors checking the validity of objects they are using, such as StackBlocksBehavior, might consider the dock object to be invalid due to it not resting flat.

This is fixed by always clamping the dock object's pose to flat