COZMO-12344 Crash in CubeAccelComponent.HandleObjectAccel (#4552)
* COZMO-12344 Crash in CubeAccelComponent.HandleObjectAccel

Was using a DEV_ASSERT to check if a pointer was null, expecting that it would never be null... apparently that is not true. Switched the DEV_ASSERT to an error. Also switched the fake cube accel console var to use LightCube numbers instead of objectID since ObjectID can change