COZMO-11225 cozmo2 no unity (#4574)
* Only deploy cozmoRobot2 if device api > 24

We now exit with a helpful error if an attempt is made to deploy the robot process
to a device that is not running at least Android 7 (sdk version 24).

* Remove Unity build dependency

- Fix `stage-assets.sh` script to use `copyResources.py` to copy assets from
  source locations

- Minor modifications to allow skipping sound assets (not currently needed for v2)

- Generate `allAssetsHash.txt` file using algorithm from csharp implemented in bash

- Skip unity build step when building cozmo v2 target from configure.py

- Add `deploy-assets.sh` script to push assets to device

  The app will search for assets in `/sdcard/Android/data/com.anki.cozmoengine/files/assets/`

- Call `stage-assets.sh` or `deploy-assets.sh` from `configure.py`, depending
  on build commands

These changes are not an ideal asset-build workflow, but are sufficient to remove
Unity as a build requirement for running Cozmov2 (Standalone). Note that deleting the
app from the phone will remove the assets. The `run-engine` script will re-deploy if necessary.
Also, if assets change in the repo and a new version of cozmoEngine standalone is installed,
the new assets may not be deployed.  In this case, call `standalone-apk/stage-assets.sh -f`
to force the new assets to be staged.

* Add script to add device IP to cozmo2Viz webots world

This automates the step of determining the IP address of a connected
android device and modifying the cozmo2Viz.wbt file to include that
info for remote visualization.