COZMO-12486 In-app text for look down laser (#4625)
* COZMO-12486 In-app text for look down laser

Changes in-app text for laser triggered by "Look Down" command to "Cozmo is looking for a laser"

* Add PounceOnMotion as part of Look Down voice command

If Cozmo does not see a laser, he switches to PounceOnMotion behavior. If he doesn't see motion then he plays an animation

Currently, there is a wasted tick between switching behaviors and we end up trying to resume a previous behavior in that tick. This results in slightly undesirable behavior like Cozmo moving his head up to try to look at a face because InteractWithFaces was resumed.