COZMO-12736 emitterutil output improvements (#4719)
* Make clad emitters generated read-only files

We currently convert CLAD emitter output to be read-only in Makefiles or other build
scripts. This change moves that logic to be a feature of `emitterutil.py` so that all
CLAD emitters that use it will output read-only files.

* Update ASTHash_emitter to use emitterutil

ASTHash_emitter is old and had its own sandboxed logic for writing output files.
We now use `emitterutils` in order to obtain standardized output.

* Ensure clad emitters output relative paths

Emitters currently accept absolute paths as inputs and output them directly into generated files.
This is potentially problematic for any kind of build artifact caching between machines in that
we might need to implement in the future.  Rather than forcing build systems to work with
relative paths, we now convert absolute paths to relative paths when generating comments for
output files.  This process now happens in `emitterutil.py`, which has access to all of the
input arguments.