Tech Debt: Fix duplicate build actions (#4718)
cozmoEngine.gyp contains two targets that both attempt to create a
symlink in (PRODUCT_DIR)/resources/config. This causes erratic build
failures when targets are built in parallel.

Fix is to add a phony target that creates the symlink, then make this a
dependency of targets that rely on it.

I applied the change to both cozmoEngine.gyp and cozmoEngine2.gyp to
keep the files in sync.  CozmoEngine2 doesn’t build recognizesFaceTool
so it doesn’t have the same problem.