Remove deprecated xctool command (#4816)
The version of xctool in our build/tools folder is not compatible with
recent versions of xcodebuild.  We don't need this tool for anything,
xcodebuild should just work.

An error in xcode.py was causing xctool to be used in the case where
xcpretty was not installed.  This change fixes the logic and removes
xctool. We no longer depend on xctool or xcpretty.