Various Voice Command related changes (#4831)
* COZMO-12862 Look Down wait time is too short

Increase look down wait for laser timeout from 3 to 5 seconds

* COZMO-12939 'Sleep Cozmo' popup although Cozmo is not on charger

We were setting a flag when we get an on charger event but not clearing it when leaving the charger

* COZMO-12143 Startled wake up not playing when cancel is pressed

Adds generic IWantsToRunStrategy, StrategyGeneric. Updates ReactionTriggerStrategyGeneric to use the new generic WantsToRun strategy.
Adds a second ReactToWakeUp behavior to the VC reaction trigger in order to play the startled wakeup animation when cancelling sleep via the cancel button. This behavior uses a generic strategy listening to the CancelIdleTimeout message