Variaous art updates (#4927)
COZMO-12965 - edited animation and art within screen to be more in line with 
  other needs views. 
COZMO-13018 - Aligning text to the left and no longer to the right. 
COZMO-11954 - Updated artwork on icons in the loader regions along bottom, 
  and updated the blue ramping up bar to no longer have Cozmo face on it and 
  just a blue gradient. Spinners also are more opaque in cube connection and 
  first connect screen. 
COZMO-12883 - Cleaned up the arrangement of text and better balanced the 
  text from art. 
COZMO-11835 - Starting on some updates with icons and colorization as well 
  as size of columns. (Still need to update the results screen without the wins, 
  and the rewards rows and icons need to be changed.), 
COZMO-12851 - Added in new art for the load bar on the loading screen. Kept
  the gradient for the fill from earlier version of app. **Also doing some updates 
  to art and placement of art within screens. Changed colors of backgrounds in 
  the games or locked states of backgrounds for Quick Tap and Memory Match.