COZMO-13355: Disable needs during Webots tests (#5000)
Our current Webots tests are not meant/designed to test needs states and it seems safer to keep them disabled by default (much like most reactions). If we want to test needs in Webots tests, we should create a test for that and specifically re-enable needs in that test's Init.