COZMO-13376 Disconnected decay bug (#5018)
We were simply pausing the entire needs system while the app was backgrounded.  Now, after
unbackgrounding, we apply decay for the period of time we were backgrounded.  In choosing the
decay config we simply look at whether we are connected to a robot at the time of unbackgrounding
and use the unconnected or the connected decay config accordingly.

- Rename ApplyDecayForUnconnectedTime -> ApplyDecayForTimeSinceLastDeviceWrite,
  and pass in a connected flag
- Call on unbackgrounding
- Adjust some comments
- Also add some 'const' to other areas