COZMO-13644 Disable Voice Commands In Engine (#5212)
This disables and removes from the build all mic capturing code and usage of speech recognition functionality for iOS and Android. For osx it is still enabled.

VoiceCommand clad definitions have been removed from the messageGameToEngine and messageEngineToGame clads, and a #define has been added in voiceCommandComponent.h that disables the broadcasting functionality.