COZMO-13551 - Code Lab: fix bug due to quotes in user project name (#5225)
In German, the "My Project 1" text is localized as "Mein Project \"1\"". The backslashes are getting removed somewhere in the process before we pass the project name to the JavaScript, which then causes the JSON parsing to barf on the multiple quotes (""Mein Project "1""). In this code change I add the backslashes back into the project names for both user and sample projects.

In addition, don't send world state (for vertical) unless we are on workspace.