Fix A/B tests initializing for robot before serial is known (#5309)
By enrolling a robot in A/B tests right when we connect to it, we
didn't yet know the serial number because the robot hadn't sent it
yet. This moves some of the A/B testing code to an area where it
can be accessed by the robot to instead initialize tests after the
robot sends us its serial number.