COZMO-13829 - Code Lab horizontal: undo removal of text_with_validator (#5334)
This was in releases Cozmo 1.6 and 1.7 for horizontal grammar only. We can't remove it because it is already in people's projects. Now it is just a copy of the 'text' block above it in the code. See text_with_validator used in below xml (from serialized project):

<block type=\"cozmo_says\" id=\"Qlgl]=Q3_CD3E=?$-U9w\">
    <value name=\"STRING\">
        <shadow type=\"text_with_validator\" id=\"CwRy%jV8C$\/;DKKBGMRZ\">
            <field name=\"TEXT\">Intruder!<\/field>
        <\/shadow>
    <\/value>
</block>

Per Gary, Logigear will test that projects created in Cozmo 1.6 and 1.7 work with Cozmo 2.0, and continue that type of testing in the future.