COZMO-13020	Rewrap - [Memory Match] How To Play screen - "You VS Cozmo"/ "Solo" buttons disappear and unable to enter game if quickly tapping Continue button at Cube setup (#5343)
When first created there was a state in memory Match setup where you would have to tap the continue button again after a scan had been stopped even when it had worked. This seemed to cause a few frames confusion if going really quickly on whether or not continue should be enabled or disabled so added some clean up to remove it completely for the faster flow.