COZMO-11225 build gardening (#5368)
* COZMO-11225 Fix paths to message-buffers in custom emitters

Add search path entry that uses the script as the base for the search
location. At least for our current code, this is a consistent way to
determine the location of `tools/message-buffers`.

* COZMO-11225 Move advertisementService into engine

AdvertisementService appear to have evolved into such a way that it now relies
on engine clad messages. These changes were made to support the SDK. Since this
service now depends on the engine, we need to move it out of coretech and into engine.
If we ever need a generic advertisementService again, we can revisit how to split
up the current functionality.

* COZMO-11225 Ignore warning for internal deprecated API

We are calling our own deprecated API. This should be removed when possible,
but for now we explicitly ignore the warning so that `-Werror` does not cause
the build to fail.

* COZMO-11225 Rename source files to reflect iososx usage

Other source files that work for both `ios` and `mac`, use the `_iososx.<ext>` pattern.

* COZMO-12960 Case-sensitive build fix

There was an additional hard-coded 'Mac' folder ref in one of the include/lib paths
in gyp.

* COZMO-11225 Fix unused var warnings

There are only a handful of unused var warnings. We should just keep the warning
flag turned on and fix them.

* Auto committing firmware from COZMO-11225_gardening