COZMO-14149 PickupObjectHelper stuck in loop (#5554)
* COZMO-14149 PickupObjectHelper stuck in loop

If the PickupObjectAction failed due to seeing an object on top of the one we are trying to pickup it would cause the PickupHelper to then try to search for the object to pickup. However, since we are already seeing the object we want to pickup, the search would immediately complete and the PickupAction would be tried again. This results in an infinite loop looking up and down the stack of objects.

This is fixed by adding a flag to PickupObjectAction's CompletionInfo that is true when the PickupObjectAction fails due to seeing an unexpected object

* Turns out this wasn't actually working because we were overriding PickupObjects completionUnion with the base class's