COZMO-11225 victor android cmake (#5571)
* COZMO-11225 xattr utility for android

This is a tiny command line utility that provides access to the android/linux
syscalls for setting/getting extended attributes on files.

We're not using APKs for deployment and instead just push build artifacts
directly to the device via ADB. The file transport is very dumb, so to avoid
unnecessary file transfers, we set an extended attribute (xattr) on each deployed
file that represents a digest of the file contents. On subsequent deploys, we
check the xattr of the digest against the actual file digest to see if we should
transfer.

This works more reliably than some of the "adb-sync" scripts floating around
that rely on filesystem time stamps. Since all of the artifacts we are deploying
are generated by the build system, we can be confident that they're not going to be
manually changed and that when something _is_ different and new file will be created.

* COZMO-11225 support ANKI_USE_JNI flag

Victor processes currently run on the "bare" android system,
outside of a typical Android "App" environment. Also, we are currently
not running any java services.  Because of this, there is no JNI environment
available, even though we're running on Android.

This change addes a new compile option to specify whether or not a
JNIenv is available, so that we can remove the assumption that JNI
will always exists on an ANDROID platform.

* COZMO-11225 robot supervisor process cleanup

from @kevinatanki

The original AndroidHAL implementation assummed we would use more of
Android's sensor APIs, which we currently do not do.

* Disabled AndroidHAL's IMU (probably won't be needed) and camera (not working yet)
* Disabled sending of AvailableAnimations as it kills the game-engine connection sometimes
* Adjusted cliff drop/undrop values with approximate thresholds
* Corrected IMU axes
* Corrected sign of motor speeds reported by syscon
* Corrected motor scaling in syscon

* COZMO-11225 cmake build support for victor/android

This change allows us to build all libraries and executables
for running victor on android with cmake. Prelim support for Mac
builds are also included, but are currently broken.

We use "modern" cmake conventions to the extent that it is possible,
and assume cmake version >= 3.6

decent overview of modern-ish cmake here: https://rix0r.nl/blog/2015/08/13/cmake-guide/

Highlights:

- clad files are now "compiled" into source files as part of the build process,
  _not_ part of a pre-build / "configure" process

- libraries that use clad depend on the actual clad files

- assets are now part of the build system in the sense that we
  use the build system to copy assets from their source locations
  in the source tree or externals dir into the build tree. This
  can be extended in the future to include processing steps.

- Nearly all adhoc "build" scripts have been eliminated and replaced
  with simpler "compiler"-pattern scripts
  (http://www.catb.org/esr/writings/taoup/html/ch11s06.html#id2958199)
  that are invoked by the build system.

- Customized android toolchain file to enable "dwo (dwarfobject)"
  file generation, C++14 and some additional optimization flags
  to speed up the build by about 5-10%.

- Cleaner library definitions and stricter include paths should
  help avoid mistakes where components are accidently used across
  library boundaries.

- One of the pain points of CMake is specifying files to compile.
  Rather than use native CMake functions for this, we use a python
  wrapper utility (metabuild.py) that interprets BUILD.in files
  similar to those used in buck and bazel and generates .lst files
  containing lists of files to build.  This provides a cleaner interface
  in cmake, but still allows for explicit auditing at the expense
  of not having _all_ info contained inside the CMakeLists.txt files.
  This also allows us to define useful custom helpers for finding and
  separating platform-specific build files.

* COZMO-11225 Victor build scripts & runtime support

These are scripts for building & deploying to android as well
as executing the programs on an android device.

build-victor.sh is the top-level build script.

Files in `scripts/` should be run from a host computer

Files in `runtime/` are copied to the android device (robot)
and are used to run programs on the device.

* COZMO-11225 Fix unused vars & symbol visibility

To trim the size of our libs & executables, we should eventually
move towards hiding all symbols by default `-Wl,-fvisibility=hidden`.
This allows for more efficient stripping, but requires that we explicitly
mark symbols that should be exported with ANKI_EXPORT.

* COZMO-11225 Run cozmoEngine as a standalone exe

This tool uses the existing CozmoAPI in the engine to run the
engine inside a standalone daemon that will run on Android.

This will need some refinement, but the basic idea is to just setup
config data, direct logs to android logcat and start the engine.
Eventually it would be good to clean this up so that it works on
mac and linux in addition to android.

* COZMO-11225 Visual Studio Code settings for victor

Visual Studio Code can be used to edit, run, deploy and debug code
to the robot. These changes include brief setup docs along with
debugger launch configurations and tasks for building/deploying/debugging
the engine running on the robot.

* COZMO-11225 Allow relative-path exclude patterns

Instead of matching only against file names, we now also match
Exclude patterns against the entire relative path. This can be used
to easily exclude entire sub-dirs. This is currently used to exclude
the `engine/tools` folder from the engine library build.

* Auto committing firmware from COZMO-11225_victor-android-cmake